+++ tools/lock/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tikkhun/lock
 
+## 0.0.7
+
+### Patch Changes
+
+- feat: 添加info输出
+
 ## 0.0.6
 
 ### Patch Changes